refactor(web): relocate Serwist provider
