+++ web/app/layout.tsx
@@ -12,9 +12,9 @@ import { ToastProvider } from './components/base/toast'
 import BrowserInitializer from './components/browser-initializer'
 import { ReactScanLoader } from './components/devtools/react-scan/loader'
 import { I18nServerProvider } from './components/provider/i18n-server'
+import { SerwistProvider } from './components/provider/serwist'
 import SentryInitializer from './components/sentry-initializer'
 import RoutePrefixHandle from './routePrefixHandle'
-import { SerwistProvider } from './serwist-provider'
 import './styles/globals.css'
 import './styles/markdown.scss'
